+++ src/Atata.HtmlValidation/GlobalSuppressions.cs
@@ -1,0 +1,7 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+#pragma warning disable S103 // Lines should not be too long
+
+[assembly: SuppressMessage("Critical Code Smell", "S2696:Instance members should not write to \"static\" fields", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.HtmlValidation.HtmlValidator.EnsureCliIsInstalled")]
+
+#pragma warning restore S103 // Lines should not be too long